Issue #4.
Updated the model to the new Outline class.  This completes the v1.0 model organization of Project -> Outline -> Tree.  Further changes are likely within these classes but the organization of these three classes is set.

This update is in preparation for implementing issue 4.  The new Outline class contains a dictionary to hold bookmarks.